# apps/api/public/openapi.yaml
openapi: 3.1.0
info:
  title: FOSL Commerce API
  version: 1.0.0
  description: |
    Public headless commerce API for the FOSL network storefront.
    Native handlers run on the API gateway (`apps/api`); checkout, orders, products,
    shipping, and webhooks are proxied to the storefront service when configured.
  contact:
    name: FOSL Support
    email: support@foslone.com
servers:
  - url: https://api.foslone.com
    description: Production
  - url: http://localhost:3002
    description: Local API gateway
...
paths:
...
  /api/v1/products:
    get:
      tags: [Products]
      summary: List products
      parameters:
        - name: scope
          in: query
          schema:
            type: string
            enum: [network, operator]
            default: network
        - name: q
          in: query
          schema:
            type: string
        - name: operatorId
          in: query
          schema:
            type: string
        - name: storefrontPath
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/StorefrontHost"
      responses:
        "200":
          description: Product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductListResponse"
...
  /api/v1/leads:
    post:
      tags: [Commerce]
      summary: Submit a lead-gen request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [productId, name, email]
              properties:
                productId:
                  type: string
                name:
                  type: string
                email:
                  type: string
                  format: email
                phone:
                  type: string
                message:
                  type: string
      responses:
        "201":
          description: Lead captured
...
  /api/v1/orders:
    post:
      tags: [Commerce]
      summary: Create order after payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrderRequest"
      responses:
        "201":
          description: Order created
    get:
      tags: [Commerce]
      summary: List orders (scoped by session or email)
      parameters:
        - name: email
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Order list
  /api/v1/orders/{id}:
    get:
      tags: [Commerce]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order detail
        "404":
          description: Not found
...
components:
  parameters:
    StorefrontHost:
      name: X-Storefront-Host
      in: header
      description: Custom domain host for CORS-scoped storefront requests
      schema:
        type: string
  schemas:
...
    ProductListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        source:
          type: string
        scope:
          type: string
...
    Product:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        type:
          type: string
          enum: [physical, digital, lead_gen]
        priceCents:
          type: integer
        vendorId:
          type: string
        vendorName:
          type: string
        category:
          type: string
        imageUrl:
          type: string
        inventory:
          type: integer
...
    CreateOrderRequest:
      type: object
      required: [email, lines]
      properties:
        email:
          type: string
        stripePaymentIntentId:
          type: string
        lines:
          type: array
          items:
            type: object
            properties:
              productId:
                type: string
              quantity:
                type: integer
              shippingMethodId:
                type: string
        shippingCents:
          type: integer
        taxCents:
          type: integer
        shipping:
          type: object
          properties:
            fullName:
              type: string
            line1:
              type: string
            city:
              type: string
            postalCode:
              type: string
            country:
              type: string